VR Hand Assets and Teleportation
I've begun the process of adding in the VR Hands and the Teleportation scripts. They are not finished and I ran into some strange errors while working on them, but I will finish them up tomorrow. But for now, these are the only changes I have made. I have also added in the OpenXR Plugin, the XR Interaction Toolkit, and Pro Builder.